+++ src/pages/MainPage.tsx
@@ -1,7 +1,13 @@
+import Calendar from '../components/Calendar';
 import Greeting from '../components/Greeting';
 
 function MainPage() {
-  return <Greeting />;
+  return (
+    <>
+      <Calendar />
+      <Greeting />
+    </>
+  );
 }
 
 export default MainPage;
